Reorder PATH variables in variables.zsh
The PATH variable "/usr/local/bin" has been moved. It was previously among in the early entries, now it has been placed at the bottom, and a note has been added to indicate it must be the first entry in the PATH. This will significantly impact the order of searching for executables in the system.

Signed-off-by: Yordis Prieto <yordis.prieto@gmail.com>