Refactor functions and update aliases in zsh configs
Removed parentheses from function declarations in the zsh configurations to adhere to a more standard syntax. A new function, `dcpsf`, was added to `functions.zsh` that formats the output of the `docker compose ps` command, and an existing alias for the same command was removed from `aliases.zsh` to avoid redundancy.

Signed-off-by: Yordis Prieto <yordis.prieto@gmail.com>